inheritance lab complete

+++ inheritance/src/main/java/inheritance/Review.java
@@ -4,7 +4,7 @@ public class Review {
 
   private String body;
   private String author;
-  private Integer numberOfStars;
+  private Double numberOfStars;
 
 
   @Override
@@ -14,7 +14,7 @@ public class Review {
         "Number of stars: " + this.getNumberOfStars();
   }
 
-  public Review(String body, String author, Integer numberOfStars) {
+  public Review(String body, String author, Double numberOfStars) {
     this.body = body;
     this.author = author;
     this.numberOfStars = numberOfStars;
@@ -36,11 +36,11 @@ public class Review {
     this.author = author;
   }
 
-  public Integer getNumberOfStars() {
+  public Double getNumberOfStars() {
     return this.numberOfStars;
   }
 
-  public void setNumberOfStars(Integer numberOfStars) {
+  public void setNumberOfStars(Double numberOfStars) {
     this.numberOfStars = numberOfStars;
   }
